requirements.txt
requests>=2.2.0
suds-jurko>=0.5
six>=1.5